+++ source/common/upstream/health_checker_base_impl.h
@@ -6,9 +6,11 @@
 #include "envoy/event/timer.h"
 #include "envoy/runtime/runtime.h"
 #include "envoy/stats/scope.h"
+#include "envoy/type/matcher/string.pb.h"
 #include "envoy/upstream/health_checker.h"
 
 #include "common/common/logger.h"
+#include "common/common/matchers.h"
 
 namespace Envoy {
 namespace Upstream {